backend/requirements.txt
fastapi>=0.115.2,<1.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.18
starlette>=0.40.0,<1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
websockets==12.0
psutil==5.9.6
GPUtil==1.4.0
pillow==10.1.0